HARP-4226: Implement raycasting of points.
