+++ @here/harp-datasource-protocol/lib/DecodedTile.ts
@@ -4,11 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as THREE from "three";
-
 import { mercatorProjection, Projection, webMercatorProjection } from "@here/harp-geoutils";
-import { HighPrecisionLineMaterial } from "@here/harp-materials";
-import { applyTechniqueToMaterial, getMaterialConstructor, Technique } from "./Techniques";
+import { Technique } from "./Techniques";
 import { TileInfo } from "./TileInfo";
 
 /**
@@ -46,12 +43,6 @@ export interface TextPathGeometry {
     technique: number;
     featureId?: number;
 }
-
-/**
- * The data stored in Buffers' elements can be of the following elementary types: float, unsigned
- * integer (either 16-bit or 32-bit long)
- */
-export type BufferElementType = "float" | "uint16" | "uint32";
 
 /**
  * Returns an array with the data type specified as parameter.
@@ -132,6 +123,22 @@ export interface Geometry {
 }
 
 /**
+ * The data stored in Buffers' elements can be of the following elementary types: float, unsigned
+ * integer (either 16-bit or 32-bit long)
+ */
+export type BufferElementType = "float" | "uint16" | "uint32";
+
+/**
+ * Structured clone compliant WebGL buffer and its metadata.
+ */
+export interface BufferAttribute {
+    name: string;
+    buffer: ArrayBufferLike;
+    type: BufferElementType;
+    itemCount: number;
+}
+
+/**
  * Structured clone compliant version of a `three.js` geometry object with text to be rendered.
  * It is composed of buffers with metadata for text objects.
  */
@@ -160,16 +167,6 @@ export interface PoiGeometry {
 }
 
 /**
- * Structured clone compliant WebGL buffer and its metadata.
- */
-export interface BufferAttribute {
-    name: string;
-    buffer: ArrayBufferLike;
-    type: BufferElementType;
-    itemCount: number;
-}
-
-/**
  * Structured clone compliant WebGL group object and its metadata.
  * Its purpose is to make working with groups of objects easier.
  */
@@ -179,99 +176,6 @@ export interface Group {
     technique: number;
     renderOrderOffset?: number;
     featureId?: number;
-}
-
-/**
- * The structure of the options to pass into [[createMaterial]].
- */
-export interface MaterialOptions {
-    /**
-     * The shader [[Technique]] to choose.
-     */
-    technique: Technique;
-
-    /**
-     * The active zoom level at material creation for zoom-dependent properties.
-     */
-    level?: number;
-
-    /**
-     * Properties to skip.
-     *
-     * @see [[applyTechniqueToMaterial]]
-     */
-    skipExtraProps?: string[];
-
-    /**
-     * `RawShaderMaterial` instances need to know about the fog at instantiation in order to avoid
-     * recompiling them manually later (ThreeJS does not update fog for `RawShaderMaterial`s).
-     */
-    fog?: boolean;
-}
-
-/**
- * Create a material, depending on the rendering technique provided in the options.
- *
- * @param options The material options the subsequent functions need.
- *
- * @returns new material instance that matches `technique.name`
- */
-export function createMaterial(options: MaterialOptions): THREE.Material | undefined {
-    const Constructor = getMaterialConstructor(options.technique);
-
-    const settings: { [key: string]: any } = {};
-
-    if (Constructor === undefined) {
-        return undefined;
-    }
-
-    if (
-        Constructor.prototype instanceof THREE.RawShaderMaterial &&
-        Constructor !== HighPrecisionLineMaterial
-    ) {
-        settings.fog = options.fog;
-    }
-
-    const material = new Constructor(settings);
-
-    if (options.technique.id !== undefined) {
-        material.name = options.technique.id;
-    }
-
-    if (options.technique.name === "extruded-polygon") {
-        material.flatShading = true;
-    }
-
-    material.depthTest =
-        options.technique.name === "extruded-polygon" && options.technique.depthTest !== false;
-
-    if (options.technique.name === "shader") {
-        // special case for ShaderTechnique.
-        // The shader technique takes the argument from its `params' member.
-        const params = options.technique.params as { [key: string]: any };
-        Object.getOwnPropertyNames(params).forEach(property => {
-            const prop = property as keyof (typeof params);
-            if (prop === "name") {
-                // skip reserved property names
-                return;
-            }
-            const m = material as any;
-            if (m[prop] instanceof THREE.Color) {
-                m[prop].set(params[prop]);
-            } else {
-                m[prop] = params[prop];
-            }
-        });
-    } else {
-        applyTechniqueToMaterial(
-            options.technique,
-            material,
-            options.level,
-            options.skipExtraProps
-        );
-    }
-
-    return material;
 }
 
 /**
@@ -302,33 +206,6 @@ export function getProjectionName(projection: Projection): string | never {
         return "webMercator";
     }
     throw new Error("Unknown projection");
-}
-
-/**
- * Returns a [[THREE.BufferAttribute]] created from a provided [[BufferAttribute]] object.
- *
- * @param attribute BufferAttribute a WebGL compliant buffer
- */
-export function getBufferAttribute(attribute: BufferAttribute): THREE.BufferAttribute {
-    switch (attribute.type) {
-        case "float":
-            return new THREE.BufferAttribute(
-                new Float32Array(attribute.buffer),
-                attribute.itemCount
-            );
-        case "uint16":
-            return new THREE.BufferAttribute(
-                new Uint16Array(attribute.buffer),
-                attribute.itemCount
-            );
-        case "uint32":
-            return new THREE.BufferAttribute(
-                new Uint32Array(attribute.buffer),
-                attribute.itemCount
-            );
-        default:
-            throw new Error(`unsupported buffer of type ${attribute.type}`);
-    } // switch
 }
 
 /**
