+++ @here/harp-datasource-protocol/package.json
@@ -1,44 +1,43 @@
 {
-  "name": "@here/harp-datasource-protocol",
-  "version": "0.2.66",
-  "description": "Datasource protocol",
-  "main": "index.js",
-  "typings": "index",
-  "directories": {
-    "test": "test"
-  },
-  "scripts": {
-    "test": "cross-env mocha --require source-map-support/register $EXTRA_MOCHA_ARGS ./test/*.js",
-    "build": "tsc",
-    "prepare": "cross-env tsc --sourceMap false $EXTRA_TSC_ARGS"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com:heremaps/harp.gl.git"
-  },
-  "author": {
-    "name": "HERE Europe B.V.",
-    "url": "https://here.com"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@here/harp-geoutils": "^0.2.40",
-    "@here/harp-lines": "^0.1.20",
-    "@here/harp-materials": "^0.1.0",
-    "@here/harp-utils": "^0.1.0"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.1.2",
-    "@types/mocha": "^5.2.5",
-    "@types/three": "0.92.24",
-    "chai": "^4.0.2",
-    "cross-env": "^5.2.0",
-    "mocha": "^5.2.0",
-    "source-map-support": "^0.5.2",
-    "three": "^0.99.0",
-    "typescript": "~3.2.1"
-  },
-  "peerDependencies": {
-    "three": "^0.99.0"
-  }
+    "name": "@here/harp-datasource-protocol",
+    "version": "0.2.66",
+    "description": "Datasource protocol",
+    "main": "index.js",
+    "typings": "index",
+    "directories": {
+        "test": "test"
+    },
+    "scripts": {
+        "test": "cross-env mocha --require source-map-support/register $EXTRA_MOCHA_ARGS ./test/*.js",
+        "build": "tsc",
+        "prepare": "cross-env tsc --sourceMap false $EXTRA_TSC_ARGS"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git://github.com:heremaps/harp.gl.git"
+    },
+    "author": {
+        "name": "HERE Europe B.V.",
+        "url": "https://here.com"
+    },
+    "license": "Apache-2.0",
+    "dependencies": {
+        "@here/harp-geoutils": "^0.2.40",
+        "@here/harp-lines": "^0.1.20",
+        "@here/harp-utils": "^0.1.0"
+    },
+    "devDependencies": {
+        "@types/chai": "^4.1.2",
+        "@types/mocha": "^5.2.5",
+        "@types/three": "0.92.24",
+        "chai": "^4.0.2",
+        "cross-env": "^5.2.0",
+        "mocha": "^5.2.0",
+        "source-map-support": "^0.5.2",
+        "three": "^0.99.0",
+        "typescript": "~3.2.1"
+    },
+    "peerDependencies": {
+        "three": "^0.99.0"
+    }
 }
