+++ @here/harp-mapview/index.ts
@@ -6,8 +6,10 @@
 
 export * from "./lib/CameraMovementDetector";
 export * from "./lib/ColorCache";
+export * from "./lib/MapViewPoints";
 export * from "./lib/composing";
 export * from "./lib/DataSource";
+export * from "./lib/DecodedTileHelpers";
 export * from "./lib/image/Image";
 export * from "./lib/image/ImageCache";
 export * from "./lib/image/MapViewImageCache";
